fix tests after new architecture
